requirements.txt
absl-py==2.1.0
accelerate==1.3.0
aiohappyeyeballs==2.3.6
aiohttp==3.10.3
aiosignal==1.3.1
# Editable install with no version control (alpha_clip==1.0)
-e /home/maiyubo/llm/AlphaCLIP-main
async-timeout==4.0.3
attrs==24.2.0
ccimport==0.4.3
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
clip @ git+https://github.com/openai/CLIP.git@dcba3cb2e2827b402d2701e7e1c7d9fed8a20ef1
contourpy==1.2.1
cumm==0.4.11
cumm-cu120==0.4.11
cycler==0.12.1
Cython==3.0.11
datasets==2.21.0
dill==0.3.8
docker-pycreds==0.4.0
einops==0.8.0
filelock==3.15.4
fire==0.6.0
fonttools==4.53.1
frozenlist==1.4.1
fsspec==2024.6.1
ftfy==6.2.3
git-lfs==1.6
gitdb==4.0.11
GitPython==3.1.43
grpcio==1.65.4
huggingface-hub==0.24.5
idna==3.7
importlib_metadata==8.2.0
importlib_resources==6.4.2
Jinja2==3.1.4
joblib==1.4.2
kiwisolver==1.4.5
lark==1.2.2
lfs==0.2
loralib==0.1.2
lvis==0.5.3
Markdown==3.6
MarkupSafe==2.1.5
matplotlib==3.9.2
mpmath==1.3.0
multidict==6.0.5
multiprocess==0.70.16
networkx==3.2.1
ninja==1.11.1.1
nltk==3.8.1
numpy==2.0.1
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.6.20
nvidia-nvtx-cu12==12.1.105
open_clip_torch==2.30.0
opencv-python==4.10.0.84
packaging==24.1
pandas==2.2.2
pccm==0.4.16
pillow==10.4.0
platformdirs==4.2.2
portalocker==2.10.1
protobuf==5.27.3
psutil==6.0.0
pyarrow==17.0.0
pybind11==2.13.6
pycocotools==2.0.8
pyparsing==3.1.2
python-dateutil==2.9.0.post0
pytz==2024.1
PyYAML==6.0.2
regex==2024.7.24
requests==2.32.3
safetensors==0.4.4
scikit-learn==1.6.1
scipy==1.13.1
sentencepiece==0.2.0
sentry-sdk==2.13.0
setproctitle==1.3.3
six==1.16.0
smmap==5.0.1
spconv-cu120==2.3.6
sympy==1.13.2
tensorboard==2.17.1
tensorboard-data-server==0.7.2
termcolor==2.4.0
threadpoolctl==3.5.0
timm==1.0.14
tokenizers==0.19.1
torch==2.4.0
torchvision==0.19.0
tqdm==4.66.5
transformers==4.44.0
triton==3.0.0
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
wandb==0.17.7
wcwidth==0.2.13
Werkzeug==3.0.3
xxhash==3.5.0
yarl==1.9.4
zipp==3.20.0
